remove class test
